Add test/slsqpcrashtest.c that reproduces a SLSQP crash.
test/slsqpcrashtest.c: New file. C reproducer for a SLSQP crash found
at DAGOPT. SLSQP does not support problems with more equality
constraints than variables, such as this one. There is supposed to be
an error code for that returned by the first invocation of slsqp by
nlopt_slsqp, but that first invocation already writes out of the bounds
of the workspace allocated by nlopt_slsqp before reporting that error,
so a segmentation fault or a "double free or corruption" usually (but
not in all contexts) happens before the error can be reported.

test/CMakeLists.txt: Build slsqpcrashtest.c.